fix a cluster rendering off by one
