+++ nats/util_test.go
@@ -15,6 +15,9 @@ package main
 
 import (
 	"testing"
+
+	"github.com/google/go-cmp/cmp"
+	"github.com/nats-io/jsm.go/api"
 )
 
 func checkErr(t *testing.T, err error, format string, a ...interface{}) {
@@ -109,5 +112,52 @@ func TestRandomString(t *testing.T) {
 			t.Fatalf("got a < 1024 length string (%d)", len(n))
 		}
 	}
+}
 
+func TestRenderCluster(t *testing.T) {
+	cluster := &api.ClusterInfo{
+		Name:   "test",
+		Leader: "S2",
+		Replicas: []*api.PeerInfo{
+			{Name: "S3", Current: false, Active: 30199700, Lag: 882130},
+			{Name: "S1", Current: false, Active: 30202300, Lag: 882354},
+		},
+	}
+
+	if result := renderCluster(cluster); result != "S1!, S2*, S3!" {
+		t.Fatalf("invalid result: %s", result)
+	}
+}
+
+func TestHostnameCompactor(t *testing.T) {
+	names := []string{
+		"broker-broker-2.broker-broker-ss.choria.svc.cluster.local",
+		"broker-broker-0.broker-broker-ss.choria.svc.cluster.local",
+		"broker-broker-1.broker-broker-ss.choria.svc.cluster.local",
+	}
+
+	result := compactStrings(names)
+	if !cmp.Equal(result, []string{"broker-broker-2", "broker-broker-0", "broker-broker-1"}) {
+		t.Fatalf("Recevied %#v", result)
+	}
+
+	names = []string{
+		"broker-broker-2.broker-broker-ss.choria.svc.cluster.local1",
+		"broker-broker-0.broker-broker-ss.choria.svc.cluster.local2",
+		"broker-broker-1.broker-broker-ss.choria.svc.cluster.local3",
+	}
+	result = compactStrings(names)
+	if !cmp.Equal(result, names) {
+		t.Fatalf("Recevied %#v", result)
+	}
+
+	names = []string{
+		"broker-broker-2.broker-broker-ss.choria.svc.cluster.local",
+		"broker-broker-0.broker-broker-ss.choria.svc.cluster.local",
+		"broker-broker-1.broker-broker-ss.other.svc.cluster.local",
+	}
+	result = compactStrings(names)
+	if !cmp.Equal(result, []string{"broker-broker-2.broker-broker-ss.choria", "broker-broker-0.broker-broker-ss.choria", "broker-broker-1.broker-broker-ss.other"}) {
+		t.Fatalf("Recevied %#v", result)
+	}
 }
